Refactor all small components #30
