+++ src/routes.tsx
@@ -1,4 +1,4 @@
-import Home from "./pages/Home";
+import Home from "./pages/home/Home";
 import Projects from "./pages/Projects";
 import React from "react";
 import {createBrowserRouter, redirect} from "react-router-dom";
